pyproject.toml
[project]
name = "smart-feeds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "google-adk>=1.18.0",
    "html2text>=2025.4.15",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "typer>=0.23.2",
]
